refactor MapViewC, ModelContainerViewC
