+++ NavioiOS/NavioiOS/MapTab/ModalContainerViewController.swift
@@ -4,14 +4,24 @@
 //
 //  Created by 구현모 on 9/11/25.
 //
-
 import UIKit
 import Combine
+import Navio
+
 
 // MARK: - ModalContainerViewController
 // 역할: 모달들의 '껍데기' 역할. 내부 컨텐츠(자식 ViewController)를 관리하고 교체
 final class ModalContainerViewController: UIViewController, UISearchBarDelegate {
-
+    
+    private let mapBoardRef: MapBoard
+    init(_ object: MapBoard) {
+        self.mapBoardRef = object
+        super.init(nibName: nil, bundle: nil)
+    }
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
     // MARK: - Properties
     // 검색창은 모든 자식 VC들이 공유해서 사용합니다
     private let searchBar: UISearchBar = {
@@ -40,6 +50,7 @@ final class ModalContainerViewController: UIViewController, UISearchBarDelegate 
         setupBinding()
         showInitialContent()
     }
+    
 
     deinit {
         cancellables.forEach { $0.cancel() }
@@ -81,7 +92,7 @@ final class ModalContainerViewController: UIViewController, UISearchBarDelegate 
     
     // 모달이 처음 나타날 때 보여줄 초기 화면 설정
     private func showInitialContent() {
-        let initialVC = LikeModalViewController()
+        let initialVC = LikeModalViewController(mapBoardRef)
         transition(to: initialVC, animated: false)
     }
